% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srpSearchSingle.json.parse.dev.r
\name{srpSearchSingle.json.parse.dev}
\alias{srpSearchSingle.json.parse.dev}
\title{Parse output from srpSearchSingle.json}
\usage{
srpSearchSingle.json.parse.dev(data)
}
\arguments{
\item{data}{Output from srpSearchSingle.json}
}
\description{
Parse output from srpSearchSingle.json
}
