% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprSearhSingle.json.parse.r
\name{srpSearchSingle.json.parse}
\alias{srpSearchSingle.json.parse}
\title{Parse output from srpSearchSingle.json.get}
\usage{
srpSearchSingle.json.parse(data)
}
\arguments{
\item{data}{Output from srpSearchSingle.json.get}
}
\description{
Parse output from srpSearchSingle.json.get
}
